requirements.txt
appdirs==1.4.4
certifi==2024.2.2
importlib-metadata==7.0.1
numpy
pandas
pyarrow==15.0.0
pyee==11.1.0
pyppeteer==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
websockets==10.4
zipp==3.17.0
